+++ tests/BaseStubsTest.php
@@ -3,7 +3,6 @@ declare(strict_types=1);
 
 namespace StubTests;
 
-use JetBrains\PhpStorm\Pure;
 use PhpParser\Node\Expr\BinaryOp\BitwiseOr;
 use PhpParser\Node\Expr\ClassConstFetch;
 use PhpParser\Node\Expr\ConstFetch;
@@ -77,7 +76,7 @@ abstract class BaseStubsTest extends TestCase
                 $value = $leftConstant->value|$rightConstant->value;
             }
         } elseif ($defaultValue instanceof UnaryMinus && property_exists($defaultValue->expr, 'value')) {
-            $value = '-' . strval($defaultValue->expr->value);
+            $value = '-' . $defaultValue->expr->value;
         } elseif ($defaultValue instanceof ClassConstFetch) {
             $class = (string)$defaultValue->class;
             if ($class === 'self' && $contextClass !== null) {
@@ -102,7 +101,6 @@ abstract class BaseStubsTest extends TestCase
         return $value;
     }
 
-    #[Pure]
     public static function getParameterRepresentation(PHPFunction $function): string
     {
         $result = '';
@@ -120,6 +118,46 @@ abstract class BaseStubsTest extends TestCase
             $result .= '$' . $parameter->name . ', ';
         }
         return rtrim($result, ', ');
+    }
+
+    /**
+     * @param array $filtered
+     * @return array
+     * @throws RuntimeException
+     */
+    protected static function getDuplicatedFunctions(array $filtered): array
+    {
+        $duplicatedFunctions = array_filter($filtered, function (PHPFunction $value, int|string $key) {
+            $duplicatesOfFunction = self::getAllDuplicatesOfFunction($value->name);
+            $functionVersions[] = Utils::getAvailableInVersions(
+                PhpStormStubsSingleton::getPhpStormStubs()->getFunction($value->name)
+            );
+            array_push($functionVersions, ...array_values(array_map(
+                fn (PHPFunction $function) => Utils::getAvailableInVersions($function),
+                $duplicatesOfFunction
+            )));
+            $hasDuplicates = false;
+            $current = array_pop($functionVersions);
+            $next = array_pop($functionVersions);
+            while ($next !== null) {
+                if (!empty(array_intersect($current, $next))) {
+                    $hasDuplicates = true;
+                }
+                $current = array_merge($current, $next);
+                $next = array_pop($functionVersions);
+            }
+            return $hasDuplicates;
+        }, ARRAY_FILTER_USE_BOTH);
+        return array_unique(array_map(fn (PHPFunction $function) => $function->name, $duplicatedFunctions));
+    }
+
+    protected static function getAllDuplicatesOfFunction(?string $name): array
+    {
+        return array_filter(
+            PhpStormStubsSingleton::getPhpStormStubs()->getFunctions(),
+            fn ($duplicateValue, $duplicateKey) => $duplicateValue->name === $name && str_contains($duplicateKey, 'duplicated'),
+            ARRAY_FILTER_USE_BOTH
+        );
     }
 
     public static function ifReflectionTypesExistInSignature(array $reflectionTypes, array $typesFromSignature): bool
