+++ tests/Parsers/ExpectedFunctionArgumentsInfo.php
@@ -3,7 +3,6 @@ declare(strict_types=1);
 
 namespace StubTests\Parsers;
 
-use JetBrains\PhpStorm\Pure;
 use PhpParser\Node\Expr;
 
 class ExpectedFunctionArgumentsInfo
@@ -64,7 +63,6 @@ class ExpectedFunctionArgumentsInfo
         return $this->index;
     }
 
-    #[Pure]
     public function __toString(): string
     {
         if ($this->functionReference === null) {
