+++ tests/StubsMetaInternalTagTest.php
@@ -5,6 +5,7 @@ namespace StubTests;
 
 use PHPUnit\Framework\Exception;
 use RuntimeException;
+use StubTests\Model\PHPFunction;
 use StubTests\Model\PHPMethod;
 use StubTests\Model\StubProblemType;
 use StubTests\Parsers\Visitors\MetaOverrideFunctionsParser;
@@ -35,6 +36,7 @@ class StubsMetaInternalTagTest extends BaseStubsTest
                     ReflectionStubsSingleton::getReflectionStubs()->getFunctions(),
                     fn ($refFunction) => $refFunction->name === $function->name
                 );
+                /** @var PHPFunction $reflectionFunction */
                 $reflectionFunction = array_pop($reflectionFunctions);
                 if (!$reflectionFunction->hasMutedProblem(StubProblemType::ABSENT_IN_META)) {
                     self::checkInternalMetaInOverride($function->name);
