+++ tests/StubsPhpDocTest.php
@@ -8,6 +8,7 @@ use phpDocumentor\Reflection\DocBlock\Tags\Link;
 use phpDocumentor\Reflection\DocBlock\Tags\Reference\Url;
 use phpDocumentor\Reflection\DocBlock\Tags\See;
 use phpDocumentor\Reflection\DocBlock\Tags\Since;
+use PHPUnit\Framework\Exception;
 use StubTests\Model\BasePHPClass;
 use StubTests\Model\BasePHPElement;
 use StubTests\Model\PHPConst;
@@ -24,6 +25,7 @@ class StubsPhpDocTest extends BaseStubsTest
      * @dataProvider \StubTests\TestData\Providers\Stubs\StubConstantsProvider::classConstantProvider
      * @param BasePHPClass $class
      * @param PHPConst $constant
+     * @throws Exception
      */
     public static function testClassConstantsPHPDocs(BasePHPClass $class, PHPConst $constant): void
     {
@@ -33,6 +35,7 @@ class StubsPhpDocTest extends BaseStubsTest
 
     /**
      * @dataProvider \StubTests\TestData\Providers\Stubs\StubConstantsProvider::globalConstantProvider
+     * @throws Exception
      */
     public static function testConstantsPHPDocs(PHPConst $constant): void
     {
@@ -42,6 +45,7 @@ class StubsPhpDocTest extends BaseStubsTest
 
     /**
      * @dataProvider \StubTests\TestData\Providers\Stubs\StubsTestDataProviders::allFunctionsProvider
+     * @throws Exception
      */
     public static function testFunctionPHPDocs(PHPFunction $function): void
     {
@@ -51,6 +55,7 @@ class StubsPhpDocTest extends BaseStubsTest
 
     /**
      * @dataProvider \StubTests\TestData\Providers\Stubs\StubsTestDataProviders::allClassesProvider
+     * @throws Exception
      */
     public static function testClassesPHPDocs(BasePHPClass $class): void
     {
@@ -60,6 +65,7 @@ class StubsPhpDocTest extends BaseStubsTest
 
     /**
      * @dataProvider \StubTests\TestData\Providers\Stubs\StubMethodsProvider::allMethodsProvider
+     * @throws Exception
      */
     public static function testMethodsPHPDocs(PHPMethod $method): void
     {
@@ -115,7 +121,7 @@ class StubsPhpDocTest extends BaseStubsTest
 
         $phpdoc = preg_replace(
             [
-                '#<br ?/>#',
+                '#<br\s*/>#',
                 '#<br>#i',
                 '#->#',
                 '#=>#',
@@ -174,6 +180,9 @@ class StubsPhpDocTest extends BaseStubsTest
         }
     }
 
+    /**
+     * @throws Exception
+     */
     private static function checkContainsOnlyValidTags(BasePHPElement $element, string $elementName): void
     {
         $VALID_TAGS = [
@@ -208,6 +217,9 @@ class StubsPhpDocTest extends BaseStubsTest
         }
     }
 
+    /**
+     * @throws Exception
+     */
     private static function checkPHPDocCorrectness(BasePHPElement $element, string $elementName): void
     {
         self::checkLinks($element, $elementName);
