+++ tests/StubsReflectionClassesTest.php
@@ -21,6 +21,7 @@ class StubsReflectionClassesTest extends BaseStubsTest
     {
         $getReturnTypeMethods = array_filter(PhpStormStubsSingleton::getPhpStormStubs()->getClass('ReflectionFunctionAbstract')
             ->methods, fn (PHPMethod $method) => $method->name === 'getReturnType');
+        /** @var PHPMethod $getReturnTypeMethod */
         $getReturnTypeMethod = array_pop($getReturnTypeMethods);
         $allReturnTypes = array_unique(Utils::flattenArray($getReturnTypeMethod->returnTypesFromAttribute +
             $getReturnTypeMethod->returnTypesFromSignature + $getReturnTypeMethod->returnTypesFromPhpDoc, false));
@@ -48,6 +49,7 @@ class StubsReflectionClassesTest extends BaseStubsTest
     {
         $getTypeMethods = array_filter(PhpStormStubsSingleton::getPhpStormStubs()->getClass('ReflectionProperty')
             ->methods, fn (PHPMethod $method) => $method->name === 'getType');
+        /** @var PHPMethod $getTypeMethod */
         $getTypeMethod = array_pop($getTypeMethods);
         $allReturnTypes = array_unique(Utils::flattenArray($getTypeMethod->returnTypesFromAttribute +
             $getTypeMethod->returnTypesFromSignature + $getTypeMethod->returnTypesFromPhpDoc, false));
@@ -70,6 +72,7 @@ class StubsReflectionClassesTest extends BaseStubsTest
     {
         $getTypeMethods = array_filter(PhpStormStubsSingleton::getPhpStormStubs()->getClass('ReflectionParameter')
             ->methods, fn (PHPMethod $method) => $method->name === 'getType');
+        /** @var PHPMethod $getTypeMethod */
         $getTypeMethod = array_pop($getTypeMethods);
         $allReturnTypes = array_unique(Utils::flattenArray($getTypeMethod->returnTypesFromAttribute +
             $getTypeMethod->returnTypesFromSignature + $getTypeMethod->returnTypesFromPhpDoc, false));
